src/main/java/sonia/scm/legacy/GravatarMD5Util.java
/*
 * Copyright (c) 2020 - present Cloudogu GmbH
 *
 * This program is free software: you can redistribute it and/or modify it under
 * the terms of the GNU Affero General Public License as published by the Free
 * Software Foundation, version 3.
 *
 * This program is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
 * details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program. If not, see https://www.gnu.org/licenses/.
 */

package sonia.scm.legacy;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.security.MessageDigest;

/**
 * Gravatar MD5 hash util from https://de.gravatar.com/site/implement/images/java/
 */
public final class GravatarMD5Util {

    private static final Logger logger = LoggerFactory.getLogger(GravatarMD5Util.class);

    private GravatarMD5Util() {
    }

    public static String md5Hex(String email) {
        try {
            MessageDigest md = MessageDigest.getInstance("MD5");
            return hex(md.digest(email.trim().toLowerCase().getBytes("CP1252")));
        } catch (Exception e) {
            logger.warn("could not create MD5 hash for email {}", email, e);
            return null;
        }
    }

    private static String hex(byte[] array) {
        StringBuilder sb = new StringBuilder(array.length * 2);
        for (byte b : array) {
            sb.append(Integer.toHexString((b & 0xFF) | 0x100).substring(1, 3));
        }
        return sb.toString();
    }
}
